refactor: clean up event filter layers
