+++ src/main/java/com/oshifes/domain/event/application/dto/EventSearchCondition.java
@@ -1,4 +1,4 @@
-package com.oshifes.domain.event.api.dto;
+package com.oshifes.domain.event.application.dto;
 
 public record EventSearchCondition(
         String country,
