--- src/main/java/com/oshifes/domain/event/entity/EventCountry.java
@@ -1,8 +1,0 @@
-package com.oshifes.domain.event.entity;
-
-public enum EventCountry {
-    KR, JP;
-
-    public static final String REGEXP = "KR|JP";
-    public static final String VALIDATION_MESSAGE = "country는 KR 또는 JP여야 합니다.";
-}
