+++ src/main/java/com/oshifes/domain/event/validation/ValidEventCountry.java
@@ -1,0 +1,21 @@
+package com.oshifes.domain.event.validation;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.FIELD, ElementType.PARAMETER})
+@Retention(RetentionPolicy.RUNTIME)
+@Constraint(validatedBy = EventCountryValidator.class)
+public @interface ValidEventCountry {
+
+    String message() default "country는 지원하는 국가 코드여야 합니다.";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+}
